# Conflicts:
#	src/main/java/com/sungjiduk/backend/common/constants/ErrorCode.java
#	src/main/java/com/sungjiduk/backend/content/repository/ContentRepository.java

+++ src/main/java/com/sungjiduk/backend/admin/dto/request/AdminSpotUpsertRequest.java
@@ -1,5 +1,7 @@
 package com.sungjiduk.backend.admin.dto.request;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
@@ -8,8 +10,8 @@ public record AdminSpotUpsertRequest(
         @NotBlank String name,
         @NotBlank String city,
         @NotBlank String address,
-        double lat,
-        double lng,
+        BigDecimal lat,
+        BigDecimal lng,
         int recommendedDurationMin,
         String referenceUrl
 ) {
